WIP: Refactoring providers into classes
